requirements.txt
google-generativeai
pypdf
reportlab
ipython